LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssScale.tentative-expected.txt


FAIL Constructing a CSSScale with an angle CSSUnitValue for the coordinates throws a TypeError assert_throws_js: function "() => new CSSScale(coord, 0)" did not throw
FAIL Constructing a CSSScale with a CSSMathValue that doesn't match <number> for the coordinates throws a TypeError assert_throws_js: function "() => new CSSScale(coord, 0)" did not throw
FAIL Constructing a CSSScale with an invalid division by px/px for the coordinates throws a TypeError assert_throws_js: function "() => new CSSScale(coord, 0)" did not throw
FAIL Updating CSSScale.x to an angle CSSUnitValue throws a TypeError assert_equals: expected "CSSUnitValue" but got "Number"
FAIL Updating CSSScale.x to a CSSMathValue that doesn't match <number> throws a TypeError assert_equals: expected "CSSUnitValue" but got "Number"
FAIL Updating CSSScale.x to an invalid division by px/px throws a TypeError assert_equals: expected "CSSUnitValue" but got "Number"
FAIL Updating CSSScale.y to an angle CSSUnitValue throws a TypeError assert_equals: expected "CSSUnitValue" but got "Number"
FAIL Updating CSSScale.y to a CSSMathValue that doesn't match <number> throws a TypeError assert_equals: expected "CSSUnitValue" but got "Number"
FAIL Updating CSSScale.y to an invalid division by px/px throws a TypeError assert_equals: expected "CSSUnitValue" but got "Number"
FAIL Updating CSSScale.z to an angle CSSUnitValue throws a TypeError assert_equals: expected "CSSUnitValue" but got "Number"
FAIL Updating CSSScale.z to a CSSMathValue that doesn't match <number> throws a TypeError assert_equals: expected "CSSUnitValue" but got "Number"
FAIL Updating CSSScale.z to an invalid division by px/px throws a TypeError assert_equals: expected "CSSUnitValue" but got "Number"
FAIL CSSScale can be constructed from two number coordinates assert_equals: expected "CSSUnitValue" but got "Number"
FAIL CSSScale can be constructed from three number coordinates assert_equals: expected "CSSUnitValue" but got "Number"
PASS CSSScale can be constructed from CSSMathValue coordinates
FAIL CSSScale.x can be updated to a number assert_equals: expected "CSSUnitValue" but got "Number"
PASS CSSScale.x can be updated to a numberish
PASS CSSScale.x can be updated to a CSSMathValue
FAIL CSSScale.y can be updated to a number assert_equals: expected "CSSUnitValue" but got "Number"
PASS CSSScale.y can be updated to a numberish
PASS CSSScale.y can be updated to a CSSMathValue
FAIL CSSScale.z can be updated to a number assert_equals: expected "CSSUnitValue" but got "Number"
PASS CSSScale.z can be updated to a numberish
PASS CSSScale.z can be updated to a CSSMathValue
PASS Modifying CSSScale.is2D can be updated to true or false
